More filters, better manual mode, tweaked filters
There's now a filter to capitalize i -> I. Currently accounts for the
edge case i.e., and nothing else. (may be expanded if necessary).

Manual mode now allows opening (for additional context), as well as the
option to allow or disallow the edit, and then open the post in a
browser. Docs have also been added.

This may need to be put on a timer in the future to make sure updates
made by dragon are included when the post edit thingy is opened.

And of course, a few filters have been made less trash.