Harden regexes, add more salutation handlers
Adds a regex to deal with "this is my first question"-meta commentary,
as well as hardening the capitalization regex. A lookahead makes sure it
matches without globbing into the match, which is apparently how non-vim
regex manages a \zs/\ze alternative.

Still primitive as fuck, but hey, it works.