Link punctuation spacing, harden help
noHelp() now tries to better respect punctuation.

By looking for the intro punctuation and the exit punctuation, a
decision is made on which punctuation to keep based on the leading
punctuation. If we have a sentence like:

    Something over 15 characters, please help me!

Dragon now preserves the ! instead of the ,.
Equivalently, if it's a !, ?, or a . instead of the `,`, that is
preserved while the (in this case) trailing ! is eradicated.

Thank fuck for replace functions :')