Add [x]: https://example.com-based links to the filter, harden regexes
Also adds some extra debugging messages to make it clear what's actually
run on which post when Dragon decides not to edit.

A few regexes have been hardened, and the capitalization regex has been
removed because regex is being a fucking cunt and refuses to let me not
match -something without needing to rewrite the regex, again.